+++ lib/address.rb
@@ -1,0 +1,13 @@
+class Address
+  def city
+    "Ростов-на-Дону"
+  end
+
+  def street
+    "ул. Маршала Конюхова"
+  end
+
+  def house
+    "д. 5"
+  end
+end
