+++ lib/order.rb
@@ -1,0 +1,9 @@
+class Order
+  def id
+    'id'
+  end
+
+  def products
+    [OpenStruct.new(weight: 20), OpenStruct.new(weight: 40)]
+  end
+end
